Created update in NurseController

+++ NursesApplication/src/main/java/com/example/nurses/NurseRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface NurseRepository extends CrudRepository<Nurse, Integer>{
+public interface NurseRepository extends CrudRepository<Nurse, Long>{
 	
 	Optional<Nurse> findByUserAndPass(String user, String pass);
 	Optional<Nurse> findByName(String name);
